Adding tokenAddress to wallet actions
