+++ server/src/utils/wallet/actions/create.js
@@ -1,25 +1,7 @@
+const { getParamsFromMethodData } = require('@utils/abi')
 const mongoose = require('mongoose')
 const WalletAction = mongoose.model('WalletAction')
-const { getParamsFromMethodData } = require('@utils/abi')
-const { get, pickBy, identity } = require('lodash')
-
-const formatActionData = ({ transactionBody, txHash, bonusType, externalId, detailedStatus, purchase }) => pickBy({
-  status: get(transactionBody, 'status'),
-  value: get(transactionBody, 'value', 0),
-  tokenName: get(transactionBody, 'tokenName', ''),
-  tokenDecimal: get(transactionBody, 'tokenDecimal', ''),
-  tokenSymbol: get(transactionBody, 'asset', ''),
-  tokenAddress: get(transactionBody, 'tokenAddress', ''),
-  from: get(transactionBody, 'from', ''),
-  to: get(transactionBody, 'to', ''),
-  blockNumber: get(transactionBody, 'blockNumber'),
-  communityName: get(transactionBody, 'communityName', ''),
-  metadata: purchase || get(transactionBody, 'tradeInfo'),
-  detailedStatus,
-  externalId,
-  bonusType,
-  txHash
-}, identity)
+const { formatActionData } = require('./utils')
 
 const handleCreateWalletJob = async (job) => {
   return new WalletAction({
@@ -54,6 +36,7 @@ const handleRelayJob = async (job) => {
         value: _amount,
         tokenAddress: _token
       },
+      tokenAddress: _token,
       walletAddress: _to,
       communityAddress: job.communityAddress || job.data.communityAddress
     }).save()
@@ -66,6 +49,7 @@ const handleRelayJob = async (job) => {
         tokenAddress: _token
       },
       walletAddress: _wallet,
+      tokenAddress: _token,
       communityAddress: job.communityAddress || job.data.communityAddress
     }).save()
     return { receiverAction, senderAction }
@@ -82,6 +66,7 @@ const handleRelayJob = async (job) => {
         value: _amount,
         tokenAddress: _token
       },
+      tokenAddress: _token,
       walletAddress: _wallet,
       communityAddress: job.communityAddress || job.data.communityAddress
     }).save()
@@ -90,10 +75,12 @@ const handleRelayJob = async (job) => {
 }
 
 const handleFundToken = (job) => {
+  const data = formatActionData(job.data)
   return new WalletAction({
     name: 'tokenBonus',
     job: mongoose.Types.ObjectId(job._id),
-    data: formatActionData(job.data),
+    data,
+    tokenAddress: data.tokenAddress,
     walletAddress: job.data.receiverAddress,
     communityAddress: job.communityAddress || job.data.communityAddress
   }).save()
@@ -140,89 +127,6 @@ const createActionFromJob = async (job) => {
   }
 }
 
-const handleSuccessCreateWalletJob = (action, job) => {
-  const formattedData = formatActionData(job.data)
-  action.set('status', get(formattedData, 'status'))
-  if (!action.walletAddress) {
-    action.set('walletAddress', job.data.walletAddress)
-  }
-  action.set('data', { ...action.data, ...formattedData })
-  return action.save()
-}
-
-const handleSuccessDefaultJob = (action, job) => {
-  const formattedData = formatActionData(job.data)
-  action.set('status', get(formattedData, 'status'))
-  action.set('data', { ...action.data, ...formattedData })
-  return action.save()
-}
-
-const handleSuccessMintDeposited = (action, job) => {
-  const formattedData = formatActionData(job.data)
-  action.set('status', get(formattedData, 'status'))
-  action.set('data', { ...action.data, ...formattedData })
-  action.set('data.detailedStatus', 'succeeded')
-  return action.save()
-}
-
-const jobSuccessHandlers = {
-  createWallet: handleSuccessCreateWalletJob,
-  createForeignWallet: handleSuccessDefaultJob,
-  fundToken: handleSuccessDefaultJob,
-  relay: handleSuccessDefaultJob,
-  mintDeposited: handleSuccessMintDeposited
-}
-
-const handlePendingAction = (action, txHash) => {
-  action.set('data.txHash', txHash)
-  return action.save()
-}
-
-const successAndUpdateByJob = async (job) => {
-  const jobHandler = jobSuccessHandlers[job.name]
-  if (!jobHandler) {
-    console.warn(`No action is defined for ${job.name}`)
-    return
-  }
-  const walletActions = await WalletAction.find({ job })
-  for (const action of walletActions) {
-    await jobHandler(action, job)
-  }
-}
-
-const pendingAndUpdateByJob = async (job, hash) => {
-  const walletActions = await WalletAction.find({ job })
-  console.log(`found ${walletActions.length} - hash ${hash}`)
-  for (const action of walletActions) {
-    await handlePendingAction(action, hash)
-  }
-}
-
-const failAndUpdateByJob = async (job) => {
-  const walletActions = await WalletAction.find({ job })
-  for (const action of walletActions) {
-    action.set('status', 'failed')
-    action.set('data', { ...action.data, ...formatActionData(job.data) })
-    action.set('failedAt', new Date())
-    action.set('failReason', job.failReason)
-    await action.save()
-  }
-}
-
-const handleSubscriptionWebHook = async (data) => {
-  return new WalletAction({
-    name: 'receiveTokens',
-    data,
-    status: 'succeeded',
-    walletAddress: data.walletAddress
-  }).save()
-}
-
 module.exports = {
-  createActionFromJob,
-  successAndUpdateByJob,
-  failAndUpdateByJob,
-  formatActionData,
-  handleSubscriptionWebHook,
-  pendingAndUpdateByJob
+  createActionFromJob
 }
